Refatorado os valores a pagar e consumir

+++ src/main/java/br/com/ficticiusclean/api/dto/PrevisaoDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class PrevisaoDTO {
 	
-	long precoGasolina;
+	double precoGasolina;
 	long totaKmPercorridoCidade; 
 	long totalKmPercorridoEstrada;
 }
